zones refactor
